% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{create_interactively}
\alias{create_interactively}
\title{Interactively Create a \code{plot2}}
\usage{
create_interactively(data = NULL)
}
\arguments{
\item{data}{A data set to load.}
}
\description{
This Shiny app allows for interactive creation of a \code{plot2}.
}
