% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_mark.R
\name{dec_mark}
\alias{dec_mark}
\alias{big_mark}
\title{Use Decimal Comma?}
\usage{
dec_mark()

big_mark()
}
\description{
These functions determine which characters the decimal mark and big mark should be that are used in plotting. They base the determination on \code{getOption("OutDec")}, which is also what \code{\link[base:format]{base::format()}} uses.
}
\details{
If the \link[=options]{option} \code{"dec_mark"} is set, that value will be used for \code{\link[=dec_mark]{dec_mark()}} if it is either a comma or a full stop.

At default, \code{\link[=big_mark]{big_mark()}} returns a full stop if \code{\link[=dec_mark]{dec_mark()}} returns a comma, and a space otherwise. If the \link[=options]{option} \code{"big_mark"} is set, that value will be used if it is either a comma (\code{","}) or a full stop (\code{"."}) or a space (\code{" "}) or an empty character (\code{""}).
}
\examples{
# at default, this follows `getOption("OutDec")`:
dec_mark()
# and big_mark() returns a space if dec_mark() returns ".":
big_mark()

# you you can set options to alter behaviour:
options(dec_mark = ",", big_mark = ".")
dec_mark()
big_mark()

options(dec_mark = ",", big_mark = NULL)
dec_mark()
big_mark()

options(big_mark = ",")
dec_mark()
big_mark()

# clean up
options(dec_mark = NULL, big_mark = NULL)
}
