% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{labellers}
\alias{labellers}
\alias{euros}
\alias{dollars}
\title{Label Euro currencies}
\usage{
euros(x, big.mark = big_mark(), decimal.mark = dec_mark(), ...)

dollars(x, big.mark = big_mark(), decimal.mark = dec_mark(), ...)
}
\arguments{
\item{x}{Values.}

\item{big.mark}{Thousands separator, defaults to \code{\link[=big_mark]{big_mark()}}.}

\item{decimal.mark}{Decimal mark, defaults to \code{\link[=dec_mark]{dec_mark()}}.}

\item{...}{Any argument to give to the geom. This will override automatically-set settings for the geom.}
}
\description{
Format numbers as currency, rounding values to dollars or cents using a convenient heuristic.
}
\examples{
\dontrun{
profit <- data.frame(group = LETTERS[1:4],
                     profit = runif(4, 10000, 25000))

profit |>
  plot2(y.labels = euros,
        datalabels = FALSE)
        
profit |>
  plot2(y.labels = euros,
        datalabels.format = euros)
}
}
