% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{plot2-options}
\alias{plot2-options}
\title{\code{plot2()} Options}
\description{
These are the package options for usage in \code{\link[=plot2]{plot2()}}. These can all be set using \code{\link[=options]{options()}}, and their values can be retrieved using \code{\link[=getOption]{getOption()}}.
}
\section{Package Options}{

\itemize{
\item \code{big_mark}\cr
Default character for a thousands separator
\item \code{dec_mark}\cr
Default character for a decimal separator
\item \code{plot2.colour}\cr
Default colour to use in plots
\item \code{plot2.colour_background}\cr
Default background colour to use in plots
\item \code{plot2.colour_font_axis}\cr
Default colour of the axis text (\code{axis.title.x} and \code{axis.title.x})
\item \code{plot2.colour_font_primary}\cr
Default colour of the plot title and plot tag
\item \code{plot2.colour_font_secondary}\cr
Default colour of the plot subtitle
\item \code{plot2.colour_sf}\cr
Default colour of the borders in sf plots
\item \code{plot2.colour_sf_fill}\cr
Default colour of the area fills in sf plots
\item \code{plot2.default_type}\cr
Default plot type
\item \code{plot2.font}\cr
Default plot font family
\item \code{plot2.silent}\cr
Default value to show console messages when plotting
\item \code{plot2.text_factor}\cr
Default text factor to use in plots
\item \code{plot2.theme}\cr
Default plot theme
}
}

