% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_pkg.R
\docType{package}
\name{plot2-package}
\alias{plot2-package}
\title{plot2: A plotting Assistant for Fast 'ggplot2' Plotting}
\description{
A streamlined extension of 'ggplot2' designed to simplify and accelerate the creation of data visualisations. 'plot2' automates common tasks such as axis handling, plot type selection, and data transformation, allowing users to create complex, publication-ready plots with minimal code. It integrates seamlessly with the tidyverse and retains full compatibility with 'ggplot2', while offering additional conveniences like enhanced sorting, faceting, and custom theming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msberends.github.io/plot2}
  \item \url{https://github.com/msberends/plot2}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.s.berends@umcg.nl}

}
\keyword{internal}
