% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_minimal2.R
\name{theme_minimal2}
\alias{theme_minimal2}
\title{An Even More Minimal Theme}
\usage{
theme_minimal2(
  ...,
  colour_font_primary = getOption("plot2.colour_font_primary", "black"),
  colour_font_secondary = getOption("plot2.colour_font_secondary", "grey35"),
  colour_font_axis = getOption("plot2.colour_font_axis", "grey25"),
  colour_background = getOption("plot2.colour_background", "white")
)
}
\arguments{
\item{...}{arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{colour_font_primary}{colour to set for the plot title and tag}

\item{colour_font_secondary}{colour to set for the plot subtitle and caption}

\item{colour_font_axis}{colour to set for the axis titles on both x and y}

\item{colour_background}{colour to set for the background}
}
\description{
This \code{ggplot2} theme provides even more white area and less clutter than \code{\link[ggplot2:ggtheme]{theme_minimal()}}.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(hp, mpg)) +
  geom_point()
  
ggplot(mtcars, aes(hp, mpg)) +
  geom_point() +
  theme_minimal2()
  
# in plot2(), the 'theme' argument defaults to theme_minimal2():
mtcars |>
  plot2(hp, mpg)
  
# set to NULL to use the ggplot2 default:
mtcars |>
  plot2(hp, mpg, theme = NULL)
}
